Move association code to assoc.rs
