+++ crates/fluvio/src/producer/dispatcher.rs
@@ -1,22 +1,22 @@
 use std::sync::Arc;
-use std::collections::{HashMap, VecDeque, HashSet};
+use std::collections::{HashMap, VecDeque};
 use futures_util::StreamExt;
 use async_channel::{Receiver, Sender};
 
 use tracing::{error, debug};
 use fluvio_types::event::StickyEvent;
-use dataplane::{ReplicaKey, Offset, ErrorCode};
-use dataplane::batch::Batch;
+use dataplane::ReplicaKey;
 use dataplane::record::{RecordSet, Record};
-use dataplane::produce::{ProduceRequest, TopicProduceData, PartitionProduceData, ProduceResponse};
+use dataplane::produce::{ProduceRequest, ProduceResponse};
 use fluvio_protocol::Encoder;
 use fluvio_protocol::api::Request;
 
 use crate::FluvioError;
 use crate::spu::SpuPool;
-use crate::producer::{Result, ProducerError};
-use crate::producer::{DispatcherMsg, ProducerConfig, ProducerMsg, PendingRecord, RecordUid};
+use crate::producer::Result;
+use crate::producer::{DispatcherMsg, ProducerConfig, ProducerMsg};
 use crate::producer::partitioning::{Partitioner, SiphashRoundRobinPartitioner};
+use crate::producer::assoc::{AssociatedResponse, AssociatedRequest, AssociatedRecord};
 
 pub(crate) struct Dispatcher {
     /// A pool of connections to the SPUs.
@@ -45,7 +45,7 @@ pub(crate) struct RecordStatusChannel {
 pub(crate) struct RecordBuffer {
     size: usize,
     max_size: usize,
-    records: VecDeque<PendingRecord>,
+    records: VecDeque<AssociatedRecord>,
 }
 
 impl RecordBuffer {
@@ -57,7 +57,7 @@ impl RecordBuffer {
         }
     }
 
-    pub fn push(&mut self, pending: PendingRecord) -> Result<(), PendingRecord> {
+    pub fn push(&mut self, pending: AssociatedRecord) -> Result<(), AssociatedRecord> {
         let record_size = Encoder::write_size(
             &pending.record,
             ProduceRequest::<RecordSet>::DEFAULT_API_VERSION,
@@ -73,7 +73,7 @@ impl RecordBuffer {
         Ok(())
     }
 
-    pub fn pop(&mut self) -> Option<PendingRecord> {
+    pub fn pop(&mut self) -> Option<AssociatedRecord> {
         let pending = self.records.pop_front();
         match pending {
             None => None,
@@ -187,7 +187,7 @@ impl Dispatcher {
 
     async fn try_buffer(
         &mut self,
-        pending: PendingRecord,
+        pending: AssociatedRecord,
         to_producer: Sender<ProducerMsg>,
     ) -> Result<(), FluvioError> {
         let batch_size = self.config.batch_size;
@@ -256,149 +256,6 @@ impl Dispatcher {
             };
         }
         Ok(())
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct BatchInfo {
-    replica_key: ReplicaKey,
-    records: HashSet<RecordUid>,
-}
-
-impl BatchInfo {
-    pub fn new(replica_key: ReplicaKey) -> Self {
-        Self {
-            replica_key,
-            records: Default::default(),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct BatchSuccess {
-    batch: BatchInfo,
-    base_offset: Offset,
-}
-
-#[derive(Debug)]
-pub(crate) struct BatchFailure {
-    batch: BatchInfo,
-    error_code: ErrorCode,
-}
-
-/// A type to build a `ProduceRequest` while associating the [`RecordUid`] of each record.
-#[derive(Debug, Default)]
-pub(crate) struct AssociatedRequest {
-    pub request: ProduceRequest,
-    pub uids: HashMap<ReplicaKey, BatchInfo>,
-}
-
-impl AssociatedRequest {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn add(&mut self, pending: PendingRecord) {
-        let topic = {
-            let maybe_topic = self
-                .request
-                .topics
-                .iter_mut()
-                .find(|t| t.name == pending.replica_key.topic);
-            match maybe_topic {
-                Some(t) => t,
-                None => {
-                    let topic = TopicProduceData {
-                        name: pending.replica_key.topic.clone(),
-                        ..Default::default()
-                    };
-                    self.request.topics.push(topic);
-                    self.request.topics.last_mut().unwrap()
-                }
-            }
-        };
-
-        let partition = {
-            let maybe_partition = topic
-                .partitions
-                .iter_mut()
-                .find(|p| p.partition_index == pending.replica_key.partition);
-            match maybe_partition {
-                Some(p) => p,
-                None => {
-                    let partition = PartitionProduceData {
-                        partition_index: pending.replica_key.partition,
-                        ..Default::default()
-                    };
-                    topic.partitions.push(partition);
-                    topic.partitions.last_mut().unwrap()
-                }
-            }
-        };
-
-        let batch = {
-            let maybe_batch = partition.records.batches.first_mut();
-            match maybe_batch {
-                Some(b) => b,
-                None => {
-                    partition.records.batches.push(Batch::new());
-                    partition.records.batches.last_mut().unwrap()
-                }
-            }
-        };
-
-        batch.add_record(pending.record);
-
-        // Associate this record's UID according to its ReplicaKey
-        let replica_key = pending.replica_key;
-        self.uids
-            .entry(replica_key.clone())
-            .or_insert_with(|| BatchInfo::new(replica_key))
-            .records
-            .insert(pending.uid);
-    }
-}
-
-pub(crate) struct AssociatedResponse {
-    successes: Vec<BatchSuccess>,
-    failures: Vec<BatchFailure>,
-}
-
-impl AssociatedResponse {
-    pub fn new(
-        response: ProduceResponse,
-        mut batches: HashMap<ReplicaKey, BatchInfo>,
-    ) -> Result<Self> {
-        let mut successes = Vec::new();
-        let mut failures = Vec::new();
-
-        for topic in response.responses {
-            for partition in topic.partitions {
-                let replica_key = ReplicaKey::new(&topic.name, partition.partition_index);
-                let batch = batches
-                    .remove(&replica_key)
-                    .ok_or_else(|| ProducerError::BatchNotFound)?;
-
-                if partition.error_code.is_ok() {
-                    let batch_success = BatchSuccess {
-                        batch,
-                        base_offset: partition.base_offset,
-                    };
-                    successes.push(batch_success);
-                } else {
-                    let batch_failure = BatchFailure {
-                        batch,
-                        error_code: partition.error_code,
-                    };
-                    failures.push(batch_failure);
-                }
-            }
-        }
-
-        Ok(Self {
-            successes,
-            failures,
-        })
     }
 }
